pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=24.1.0",
    "aiohttp>=3.12.15",
    "asyncio-throttle>=1.0.2",
    "asyncpg>=0.30.0",
    "beautifulsoup4>=4.13.5",
    "facebook-sdk>=3.1.0",
    "fake-useragent>=2.2.0",
    "fastapi>=0.116.1",
    "feedparser>=6.0.12",
    "groq>=0.31.1",
    "lxml>=6.0.1",
    "nltk>=3.9.1",
    "pillow>=11.3.0",
    "playwright>=1.55.0",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "python-telegram-bot>=22.3",
    "pyyaml>=6.0.2",
    "redis>=6.4.0",
    "requests>=2.32.5",
    "schedule>=1.2.2",
    "selenium-stealth>=1.0.6",
    "textstat>=0.7.10",
    "tiktoken>=0.11.0",
    "uvicorn>=0.35.0",
]
